Relocate IUndoManager + undo history types into headless Gum.Presentation (Phase 3, ADR-0005)
Continues ADR-0005 Phase 3 (after #3345/#3346/#3347). Moves the IUndoManager contract and the framework-neutral undo data-type cluster into the headless Gum.Presentation assembly, preserving the Gum.Undo namespace so consumers are unchanged. The concrete UndoManager implementation stays in the WPF shell.

Relocated into Tools/Gum.Presentation/Undo/: IUndoManager (git rename), all of UndoSnapshot.cs (UndoComparison + UndoSnapshot, git rename), and the data cluster extracted out of UndoManager.cs into new files -- UndoLock.cs, ElementHistory.cs (HistoryAction + ElementHistory), BehaviorHistory.cs (BehaviorSnapshot + BehaviorHistoryAction + BehaviorHistory), and UndoOperation.cs (enum + UndoOperationEventArgs). All depend only on GumCommon types, so they compile headless.

Back-edge break: UndoLock.Dispose previously called _undoManager.UndoLocks.Remove(this), a direct reference back to the shell-resident UndoManager. UndoLock now takes an injected Action onDispose, and UndoManager.RequestLock wires it as () => UndoLocks.Remove(theLock). Behavior-preserving; pinned by a new test (RequestLock -> Dispose still removes the lock) plus a headless UndoLock dispose-callback test.

Also drops 3 dead 'using System.Windows.*' lines on the way (UndoSnapshot.cs x2, UndoManager.cs x1).

Proof: Gum.Presentation builds standalone (0 errors, no WPF/WinForms reference); GumFull.sln builds (0 errors); GumToolUnitTests green (998 passed, 4 skipped); Gum.Presentation.Tests green (4 passed).
