+++ Tools/Gum.Presentation/Undo/IUndoManager.cs
