+++ Tools/Gum.Presentation/Undo/UndoSnapshot.cs
@@ -5,8 +5,6 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
-using System.Windows.Controls;
-using System.Windows.Documents;
 
 namespace Gum.Undo;
